requirements.txt
virtualenv
python-telegram-bot

# Linting dependencies
pylint